refactor(backend): organize modules by architectural responsibility
- move HTTP handlers, models, errors, and rate limiting into `api`
- group application state and use-case services under `application`
- relocate business roles and intake assessment into `domain`
- isolate AI and map adapters under `integrations`
- move SeaORM entities into `persistence`
- preserve legacy public module paths through crate re-exports
- update imports in configuration and server bootstrap
- remove the obsolete architecture document
