--- docs/ARCHITECTURE.md
@@ -1,166 +1,0 @@
-# 技术架构
-
-## 1. 状态与原则
-
-本文同时记录已经落地的工程基线与尚未实现的候选业务架构。工程落地时应优先保证可演示、可审计和可收缩，而不是追求无人监督的复杂调度。
-
-MVP 以人工在环为基本原则。后续版本可以提高信息整理、风险提示、低风险任务编排和流程提醒的自动化程度，但事实确认、案情发布、权限变更和高风险任务派遣必须保留人工审批。任何自动化能力都必须支持暂停、撤回、审计和人工接管。
-
-当前已确定并实现的工程基线：
-
-- 单仓库结构，Rust 后端位于仓库根目录，React 前端单独位于 `frontend/`。
-- 前端使用 React 19、TypeScript、Vite、HeroUI 3、Tailwind CSS 4、React Router 和 Lucide 图标。
-- 后端使用 Rust 2024、Actix Web、Actix CORS、Serde 和 `env_logger`。
-- 数据层使用 SeaORM 1.1 稳定版本线，启用 SQLite、PostgreSQL 和 MySQL 驱动；迁移由独立的 `migration/` workspace crate 管理。
-- Vite 开发服务器将 `/api` 代理到 Actix 默认地址 `127.0.0.1:8080`。
-- 已实现健康检查、案件创建/列表/详情、案件状态流转、线索提交和线索人工审核 API。
-- 案件、老人画像、线索和审计事件已建立 SeaORM Entity；关键写入与审计事件在同一数据库事务内提交。
-- SQLite 已完成迁移 `up -> down -> up`、业务事务测试和 HTTP 闭环联调。
-- 身份使用 Argon2id 密码哈希和数据库短期会话；Bearer 原文不入库，服务端只保存 SHA-256 哈希并支持撤销。
-- `case_memberships` 同时约束案件可见性和操作权限；家属、指挥、志愿者的线索与健康字段视图由服务端生成。
-
-密码找回、MFA、账号审批、组织模型、任务、轨迹、AI、地图和文件模块仍属于计划架构。
-
-核心原则：
-
-- 人工在环：AI 生成草稿，授权人员决定事实、任务和公开内容。
-- 最小权限：各端仅获得完成当前职责所需的数据和操作。
-- 状态可追溯：线索、摘要、任务和权限变更保留操作者、时间和前后值。
-- 数据分级：公开、协作、内部、敏感数据在接口和存储层隔离。
-- 模拟优先：MVP 默认使用虚构或脱敏数据，真实接入另行评审。
-- 可替换集成：AI、地图、存储等外部服务通过适配层接入，避免业务逻辑与单一供应商 SDK 紧耦合。
-
-## 2. 候选系统结构
-
-```text
-家属端 ─────┐
-指挥端 ─────┼──> API / 身份与权限 ──> 案件服务
-志愿者端 ───┘          │              ├─ 线索与审核
-                       │              ├─ 任务与轨迹
-                       │              ├─ 案情视图
-                       │              └─ 审计日志
-                       │
-                       ├──> AI 编排层 ──> 通义千问/百炼/RAG
-                       ├──> 地图适配层 ─> 高德地图 API
-                       └──> 文件适配层 ─> 本地/对象存储
-```
-
-建议把面向不同角色的“可见视图”放在服务端生成，不能只依赖前端隐藏字段。
-
-## 3. 模块边界
-
-### 3.1 身份与权限
-
-- 已实现用户、全局角色、账号状态、数据库会话和案件成员关系。
-- 已实现登录、登出、会话过期/撤销、进程内登录失败限流和服务端案件授权检查。
-- 家属创建案件时只成为该案件的家属成员，可按邮箱显式邀请全局角色匹配的指挥账号；指挥成为成员后可继续添加其他角色匹配的案件成员。
-- 管理员角色不自动获得业务案件访问权。
-- 演示账号使用 `.invalid` 保留域，并通过显式 CLI 初始化；应用启动不会自动创建账号。
-- 账号审批、密码找回、MFA、组织/队伍边界和高风险二次确认仍待实现。
-
-### 3.2 案件服务
-
-- 案件基本信息、状态和参与人员。
-- 老人画像的版本、确认状态和敏感字段。
-- 不同角色的案件摘要视图。
-- 案件结束、冻结、归档和数据保留策略。
-
-### 3.3 线索与审核
-
-- 保存原始输入引用、AI 提取草稿和人工审核结果。
-- 线索的来源、时间、地点、可信度依据和状态流转。
-- 避免覆盖历史值；更正应产生新版本和审计记录。
-
-推荐状态流：
-
-```text
-待处理 -> 待核实 -> 已确认
-   │         ├──> 已排除
-   │         └──> 已失效
-   └────────────> 重复/无关
-```
-
-### 3.4 任务与轨迹
-
-- 任务位置、目标、负责人、截止时间、风险提示和来源线索。
-- 接受、开始、暂停、完成、取消与异常状态。
-- 位置回传只在明确参与任务期间启用，并支持主动停止。
-- 原始轨迹与面向界面的简化轨迹分开处理。
-
-### 3.5 AI 编排层
-
-- 为问询、线索结构化、案情总结、知识问答和案例整理提供独立任务模板。
-- 在调用模型前进行权限检查、字段裁剪和必要脱敏。
-- 保存提示模板版本、模型标识、输入引用、输出和审核结果，敏感内容按策略脱敏或加密。
-- 使用结构化输出校验；解析失败时进入人工处理，不静默猜测字段。
-- 对提示注入、恶意附件和知识库污染设置防护与测试。
-
-### 3.6 地图适配层
-
-- 坐标转换、地理编码、POI、路线、天气/路况和地图展示数据。
-- 业务层保存标准化位置及来源，不直接依赖页面 SDK 的临时对象。
-- 对不同角色输出不同精度的位置；非必要场景使用模糊化位置。
-- 外部 API 失败时允许退化为静态点位或手工任务，不能阻塞安全相关操作。
-
-### 3.7 知识库与案例库
-
-- 只接收来源明确、获得授权、完成脱敏并通过审核的内容。
-- 文档保存版本、适用地区、有效期和审核人。
-- 检索结果需携带来源，无法找到可靠依据时应明确回答“不确定/需要咨询负责人”。
-
-## 4. 数据存储基线
-
-数据库层已落地首批案件纵向闭环：
-
-- ORM：统一使用 SeaORM，不在业务服务中直接拼接 SQL。
-- 数据库兼容：同时支持 SQLite、PostgreSQL 和 MySQL；本地开发默认 SQLite，部署环境可按需要选择 PostgreSQL 或 MySQL。
-- 迁移框架：使用 `sea-orm-migration` 管理迁移状态和执行顺序。
-- SQL 留档：每次结构变化必须保留“全局编号 + 功能名”的 `.sql` 文件，例如 `0001_create_cases.sql`。
-- 方言隔离：SQLite、PostgreSQL、MySQL 使用相同编号和功能名，分别存放在各自目录中，不能用运行时字符串替换拼装方言 SQL。
-- 不可变更：迁移合并后不得重排编号、删除旧脚本或直接修改已经发布的脚本；修正必须新增迁移编号。
-- 禁止自动同步：应用启动时不得使用自动建表或 schema synchronize 替代版本化迁移。
-- 首批表：`cases`、`elder_profiles`、`clues`、`audit_events`。
-- 身份权限表：`users`、`auth_sessions`、`case_memberships`、`clue_attributions`。
-- 标识与时间：首版跨数据库基线使用字符串 UUID 和 RFC3339 UTC 时间，避免业务 Entity 被单一数据库原生 UUID/时间类型绑定；未来改为原生类型必须新增迁移。
-- 事务边界：`case.created`、`case.status_changed`、`clue.submitted`、`clue.reviewed` 与对应业务写入原子提交。
-- 文件：开发阶段使用受控本地目录；部署时评估带访问控制和生命周期规则的对象存储。
-- 缓存/任务队列：首版可不引入，确有异步 AI、附件处理或通知需求时再增加。
-- 向量检索：优先使用百炼知识库或独立适配的本地向量库，避免成为案件事实的唯一存储。
-
-完整的目录、命名、回滚和验证规则见 [DATABASE.md](./DATABASE.md)。
-
-PostgreSQL/MySQL 已完成依赖编译和方言 SQL 留档，但仍需在 CI 服务容器或真实测试实例中执行完整迁移与业务集成测试，不能以 SQLite 结果替代。
-
-敏感字段的加密、密钥托管、备份与删除能力必须在接入真实数据前完成。
-
-## 5. 关键接口约束
-
-接口设计至少遵守以下规则：
-
-- 所有案件资源都验证用户与案件的关系，不仅验证是否登录。
-- 列表和详情接口执行字段级过滤，不能返回后再由前端删除。
-- 上传接口限制类型、大小和数量，进行恶意文件检查，并移除不必要的图片元数据。
-- 位置回传包含采集时间、精度、来源和任务关联，拒绝明显过期或格式异常的数据。
-- AI 结果必须带状态（草稿/已审核/已拒绝）、模型与模板版本。
-- 修改事实、发布摘要、分配任务、导出和权限变更均产生审计事件。
-- API 日志不得记录令牌、完整联系方式、病史、原始聊天全文或精确轨迹。
-
-## 6. 环境建议
-
-至少区分：
-
-- `development`：本地开发，仅使用模拟数据和开发密钥。
-- `demo`：比赛/评审演示，使用固定脱敏数据，限制外部写入。
-- `staging`：集成与安全测试，不复制生产敏感数据。
-- `production`：只有完成合规、安全、运维和应急评审后才能建立。
-
-环境配置通过变量或密钥服务注入。仓库只提供去敏后的 `.env.example`，不得提供任何可用密钥。
-
-## 7. 质量验证建议
-
-- 单元测试：权限规则、状态机、脱敏器、结构化输出校验。
-- 集成测试：AI/地图失败与超时、文件上传、数据库事务。
-- 端到端测试：三端完整模拟案件与越权场景。
-- 安全测试：水平/垂直越权、ID 枚举、提示注入、恶意附件、敏感日志。
-- AI 评估：事实一致性、遗漏率、错误确定性、引用正确性、不同角色的信息泄露。
-- 演练：地图或 AI 服务不可用、志愿者停止回传、误发线索和账号泄露。
