+++ src/api/mod.rs
@@ -1,0 +1,5 @@
+pub mod auth;
+pub mod error;
+pub mod models;
+pub mod rate_limit;
+pub mod routes;
