+++ src/api/routes/health.rs
