+++ src/application/services/auth_service.rs
