+++ src/application/services/case_resource_service.rs
