+++ src/application/services/case_service.rs
