+++ src/domain/mod.rs
@@ -1,0 +1,2 @@
+pub mod intake_assessment;
+pub mod roles;
