+++ src/integrations/ai_gateway.rs
