+++ src/integrations/mod.rs
@@ -1,0 +1,2 @@
+pub mod ai_gateway;
+pub mod amap_service;
