+++ src/persistence/entities/ai_prompt_templates.rs
