+++ src/persistence/entities/case_places.rs
