+++ src/persistence/entities/cases.rs
