+++ src/persistence/entities/task_location_reports.rs
