+++ src/persistence/entities/user_global_capabilities.rs
